Move notifications feature into 'Feature' folder
