--- src/app/core/notification/components/dialog-info/dialog-info.component.html
@@ -1,6 +1,0 @@
-<div style="display: flex; max-width: 300px" fxLayout="column" fxLayoutGap="15px">
-    <div fxFlex style="justify-content: center;" fxFlexAlign="center center">{{data}}</div>
-    <div fxFlex fxLayoutAlign="end">
-        <button mat-button (click)="onCloseClick()">Close</button>
-    </div>
-</div>
