+++ src/app/features/notification/components/dialog-info/dialog-info.component.html
@@ -1,0 +1,6 @@
+<div style="max-width: 400px">
+    <div>{{data}}</div>
+    <div style="margin-top:30px;">
+        <button mat-button (click)="onCloseClick()">Close</button>
+    </div>
+</div>
